move RepositoryGrid components into own folder
